Created more core events.
pubsub is definitely the way to go with Backbone.js (with bind/trigger
being the Backbone.js nomenclature) and I've altered a few methods and
converted them to methods bound to the rendering rendered view events.

Additionally, I'll no longer have jQuery add classes with duration. CSS3
Animation is probably the best way to take it. Older clients aren't
great with jQuery's addClass animation so it's probably better to have
it only occur on devices with CSS3 Animation.