Initial refactoring:
Refactoring the constructor functions with a class method approach to
allow reuse of any useful functionality and to allow future testing.

As I've updated Backbone, I've renamed the (potentially going to be
deprecated) bind methods to on as it reduces confusion as I'm also using
Underscore.

Backbone now has cached jQuery objects in views and I've slowly moved my
own versions of this (which I often did on a method-by-method basis) to
the Backbone variables.